+++ tests/test_filelock.py
@@ -371,6 +371,21 @@ def test_poll_intervall_deprecated(lock_type: type[BaseFileLock], tmp_path: Path
             pytest.fail("No warnings of stacklevel=2 matching.")
 
 
+
+@pytest.mark.parametrize("lock_type", [FileLock, SoftFileLock])
+def test_context_decorator(lock_type: type[BaseFileLock], tmp_path: Path) -> None:
+    lock_path = tmp_path / "a"
+    lock = lock_type(str(lock_path))
+
+    @lock
+    def decorated_method() -> None:
+        assert lock.is_locked
+
+    assert not lock.is_locked
+    decorated_method()
+    assert not lock.is_locked
+
+
 @pytest.mark.skipif(sys.platform != "win32", reason="Tests behavior of UnixFileLock on Windows systems")
 def test_unix_lock_on_windows(tmp_path: Path) -> None:
     lock = UnixFileLock(str(tmp_path / "lockfile"))